refactor(rsc): extract server references manager
